Backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
sentence-transformers
transformers[torch] # or [tf] if using TensorFlow
pinecone # Ensure you have the client library (v3+)
pillow
requests
joblib
langchain
langchain-community
python-dotenv # Good practice for managing API keys